requirements.txt
numpy
nltk
scikit-learn
inltk
indic-nlp-library
matplotlib